Move color properties (color index, direction index) into separate class.
